backend/requirements.txt
# FastAPI и сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.3.0

# Authentication (базовая JWT для MVP)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI Services
openai==1.3.5
pinecone-client==2.2.4  # Опционально, если хотите использовать Pinecone
qdrant-client==1.7.0  # Векторная БД (замена Pinecone)

# HTTP клиенты
httpx==0.25.1
requests==2.31.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Media processing
Pillow==10.1.0

# AWS S3
boto3==1.29.7
botocore==1.32.7

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

